backend/admin.py
from django.conf.locale.en import formats as en_formats
en_formats.DATE_FORMAT = "Y-m-d"

from django.contrib import admin
from django.contrib.auth import get_user_model
from django.utils.translation import ugettext_lazy as _
from django.db import models

from django_json_widget.widgets import JSONEditorWidget
from adminsortable2.admin import SortableAdminMixin
